Separate fc_core.c from RC processing
